vello_cpu::coarse

Module bucketer

Structs

ActiveLayer 🔒

Metadata about the currently active layer.

CommandBucketer 🔒

A bucketer that groups commands into strip-row-sized buckets.

GeneratedAlphaFill 🔒

A generic alpha fill to allow using generate_fill to create either paint fills or blend fills.

GeneratedFill 🔒

A generic fill to allow using generate_fill to create either paint fills or blend fills.

RowLayerState 🔒

Each recorded layer by itself already stores a bounding box. However, we want to take this one step further and track a horizontal bounding box of each row band individually. This allows us to save a lot of work when clearing buffers during fine rasterization.

RowState 🔒

State for a single row of strips.

Functions

debug_assert_tile_aligned 🔒