vello_cpu::coarse

Module cmd

Render commands for a single row of strips.

Structs

AlphaIdx 🔒

DepthFill 🔒

Fill a whole range of depth buckets with the given paint.

LayerFill 🔒

Composite a span from the current temporary layer buffer into the parent buffer and optionally apply some alpha coverage.

LayerFillAttrs 🔒

PaintFill 🔒

Fill a span with the given paint and optionally some alpha coverage.

Enums

RenderCmd 🔒

A bucketed render command.