vello_cpu::dispatch

Module multi_threaded

Modules

cost 🔒

worker 🔒

Structs

AllocationGroup 🔒

AllocationManager 🔒

A structure that allows storing and fetching existing allocations.

Allocations 🔒

A structure to keep track of allocations that will be done while rendering with multi-threading.

MaybePresent 🔒

An object that might hold a certain value (behind a mutex), and panics if we attempt to access it when it’s not initialized.

MultiThreadedDispatcher 🔒

A dispatcher for multi-threaded rendering.

OwnedClip 🔒

RecordedCommandTask 🔒

RenderTask 🔒

Enums

RecordedCommand 🔒

RenderTaskType 🔒

Type Aliases

RecordedCommandReceiver 🔒

RecordedCommandSender 🔒

RenderTaskSender 🔒