vello_cpu

Module render

Basic render operations.

Structs

ImageRegistry 🔒

Registry that maps opaque ImageIds to Pixmap data.

RasterizerSettings

Settings used when rasterizing a scene into a pixmap.

RenderContext

A render context for CPU-based 2D graphics rendering.

RenderSettings

Settings to apply to the render context.

Resources

Persistent resources required by Vello CPU for rendering.

Enums

CompositeMode

The composition mode that should be used when rendering into a pixmap.

PixelFormat

The pixel format to assume for the destination pixmap.

Constants

ATLAS_IMAGE_ID_BASE 🔒

DEFAULT_GLYPH_ATLAS_SIZE 🔒