vello_cpu

Struct Pixmap

pub struct Pixmap {
    width: u16,
    height: u16,
    buf: Vec<PremulRgba8>,
    may_have_transparency: bool,
}

A pixmap of premultiplied RGBA8 values backed by u8.

Fields
width: u16height: u16buf: Vec<PremulRgba8>may_have_transparency: bool
Implementations

impl Pixmap

pub fn new(width: u16, height: u16) -> Pixmap

Create a new pixmap with the given width and height in pixels.

All pixels are initialized to transparent black.

pub fn from_parts(data: Vec<PremulRgba8>, width: u16, height: u16) -> Pixmap

Create a new pixmap with the given premultiplied RGBA8 data.

The data vector must be of length width * height exactly.

The pixels are in row-major order.

This assumes the image may have transparent pixels. Use from_parts_with_opacity if you already know the opacity status to enable optimizations.

Panics

Panics if the data vector is not of length width * height.

pub fn from_parts_with_opacity( data: Vec<PremulRgba8>, width: u16, height: u16, may_have_transparency: bool, ) -> Pixmap

Create a new pixmap with the given premultiplied RGBA8 data and precomputed opacity flag.

The data vector must be of length width * height exactly.

The pixels are in row-major order.

Use this when you’ve already determined whether the data contains non-opaque pixels to avoid redundant scanning.

Panics

Panics if the data vector is not of length width * height.

pub fn resize(&mut self, width: u16, height: u16)

Resizes the pixmap container to the given width and height; this does not resize the contained image.

If the pixmap buffer has to grow to fit the new size, those pixels are set to transparent black. If the pixmap buffer is larger than required, the buffer is truncated and its reserved capacity is unchanged.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the pixmap buffer to fit the pixmap’s current size.

pub fn capacity(&self) -> usize

The reserved capacity (in pixels) of this pixmap.

When calling Pixmap::resize with a width * height smaller than this value, the pixmap does not need to reallocate.

pub fn width(&self) -> u16

Return the width of the pixmap.

pub fn height(&self) -> u16

Return the height of the pixmap.

pub fn may_have_transparency(&self) -> bool

Returns whether the pixmap may have non-opaque pixels.

This value is computed at construction time. It may become stale if pixels are modified directly via data_mut(), data_as_u8_slice_mut(), or set_pixel().

Use set_may_have_transparency() to manually update the flag, or recompute_may_have_transparency() to recalculate it by scanning all pixels.

pub fn set_may_have_transparency(&mut self, may_have_transparency: bool)

Manually set the may_have_transparency flag.

Use this after modifying pixels via data_mut() or set_pixel() when you know whether the image has non-opaque pixels.

pub fn recompute_may_have_transparency(&mut self)

Recalculate may_have_transparency by scanning all pixels.

Use this after modifying pixels via data_mut() or set_pixel() when you need accurate opacity information.

pub fn multiply_alpha(&mut self, alpha: u8)

Apply an alpha value to the whole pixmap.

pub fn from_png(data: impl BufRead + Seek) -> Result<Pixmap, DecodingError>

Create a pixmap from a PNG file.

pub fn into_png(self) -> Result<Vec<u8>, EncodingError>

Return the current content of the pixmap as a PNG.

pub fn data(&self) -> &[PremulRgba8]

Returns a reference to the underlying data as premultiplied RGBA8.

The pixels are in row-major order.

pub fn data_mut(&mut self) -> &mut [PremulRgba8]

Returns a mutable reference to the underlying data as premultiplied RGBA8.

The pixels are in row-major order.

pub fn data_as_u8_slice(&self) -> &[u8]

Returns a reference to the underlying data as premultiplied RGBA8.

The pixels are in row-major order. Each pixel consists of four bytes in the order [r, g, b, a].

pub fn data_as_u8_slice_mut(&mut self) -> &mut [u8]

Returns a mutable reference to the underlying data as premultiplied RGBA8.

The pixels are in row-major order. Each pixel consists of four bytes in the order [r, g, b, a].

pub fn as_mut(&mut self) -> PixmapMut<'_>

Return a mutable view into this pixmap’s pixel data.

pub fn sample(&self, x: u16, y: u16) -> PremulRgba8

Sample a pixel from the pixmap.

The pixel data is premultiplied RGBA8.

pub fn sample_idx(&self, idx: u32) -> PremulRgba8

Sample a pixel from a custom-calculated index. This index should be calculated assuming that the data is stored in row-major order.

pub fn set_pixel(&mut self, x: u16, y: u16, pixel: PremulRgba8)

Set a pixel in the pixmap at the given coordinates.

The pixel data should be premultiplied RGBA8. The coordinate system has its origin at the top-left corner, with x increasing to the right and y increasing downward.

pub fn take(self) -> Vec<PremulRgba8>

Consume the pixmap, returning the data as the underlying Vec of premultiplied RGBA8.

The pixels are in row-major order.

pub fn take_unpremultiplied(self) -> Vec<Rgba8>

Consume the pixmap, returning the data as (unpremultiplied) RGBA8.

Not fast, but useful for saving to PNG etc.

The pixels are in row-major order.